`sp_remove` catches `ValueError` if `some_path` is not in `sys.path` (#9)
Function `sp_remove` uses a `try` block rather than verifying the
argument's absence with an `if` block. If the bad situation occurs,
`try` is faster than `if`. In most cases, presumably, the users will
attempt to remove paths that they have added to `sys.path`.

The wildcard import excludes the packages' modules: they are not
supposed to be directly used out of their package.

A warning not specific to `syspathmodif` was removed from `README.md`.

`syspathmodif`'s version was incremented to `1.1.2`.